end of day commit

+++ src/ludumDare/game/entity/Entity.java
@@ -23,7 +23,7 @@ public abstract class Entity {
 		return null;
 	}
 	
-	public Entity(int x, int y) {
+	public Entity(float x, float y) {
 		this.x = x;
 		this.y = y;
 	}
